Engine Overhaul - Version 1.1.8 BETA
Completely changed Andromeda engine to mimic Godot in structure.
Changed the node tree system such that it is almost exactly like Godot.
Base node is 'node', with 'CanvasLayer' and 'Sprite' branching from it.
Sprites are the basae node for almost all game objects.

The levels are saved in a builder format, utilizing scenes, which are
just predefined sets of nodes that can take in some input to define its
properties.

Reintroduced the seperate window logging system.

Changed the image loading system to be more generic and to utilize
ImageBitmaps for faster rendering.

Slightly tweaked the platform textures and completely reworked the
player texture.