Add more relic types
Add the feather, battery, and ancient shield. The feather will
make the player jump higher for one jump. That jump is activated when
the feather is used, and can be used once while in the air, allowing for
a double jump.
The battery allows the player to reverse gravity for a maximum of 5
seconds, but can be stopped early by activating the relic again. The
relic count will only decrease when the gravity switches back to normal.
The ancient shield makes the player invincible for 5 seconds.